refactor/EV-739/refactoring widgets
