--- lib/bottom_navigation_icons.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class BottomNavigationIcons {
-  BottomNavigationIcons._();
-
-  static const _kFontFam = 'BottomNavigationIcons';
-  static const String? _kFontPkg = null;
-
-  static const IconData balance =
-      IconData(0xe802, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData wallet =
-      IconData(0xe803, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData warren =
-      IconData(0xe804, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-}
