+++ lib/utils/providers/providers.dart
@@ -1,0 +1,5 @@
+import 'package:riverpod/riverpod.dart';
+
+final visibilityProvider = StateProvider<bool>(
+  (ref) => true,
+);
